Update script.js
## Changes Made to Cross-site Search Functionality

### 1. `index.html`:

**- Added JavaScript Link**:
Added a `<script>` link to load the `script.js` file. This will help run our custom JavaScript functions for the search functionality.

**- Checkbox Event Listener**:
Included an `onclick` attribute to the 'cross-site-search' checkbox to call a function `deselectOthers()`. This function ensures that if 'All Sites' is selected, other individual checkboxes will be deselected.

### 2. `script.js`:

**- Array of Sites**:
The `sites` array was uncommented to store the list of sites from the checkboxes.

**- Search Functionality**:
The search button now listens for a click event. On click, it will:

- Check if the 'All Sites' checkbox is selected. If yes, all sites will be used for cross-site search.
- If 'All Sites' is not selected, only those sites for which the checkboxes are checked will be used.
- Constructs a search query for Google using the `site:` operator to narrow down the results.
- Opens the Google search in a new tab.

**- Open in New Tab Functionality**:
Added a function `openNewTab(url)` to help open URLs in a new browser tab.

**- Deselecting Other Checkboxes**:
When the 'All Sites' checkbox is checked, all other checkboxes will be deselected to ensure clarity for the user.

---

### To implement:

- Simply replace the old `index.html` and `script.js` files with the newly provided versions.
- Ensure that both files are in the same directory for the JavaScript to correctly interact with the HTML.